Moves test RPC client to test folder.
